# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MERNSecrets API
  version: 0.1.0
  description: Authentication, organization, project, environment, secret, audit, and RBAC API.
servers:
  - url: http://localhost:4000/api
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    Error:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: Forbidden }
  responses:
    BadRequest:
      description: Invalid request data
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Authentication required or invalid
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Permission denied
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Conflict:
      description: Duplicate resource
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
paths:
...
  /projects:
    get: { security: [{ accessCookie: [] }], parameters: [{ name: organizationId, in: query, required: true, schema: { type: string } }], responses: { '200': { description: Project list }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } } }
    post: { security: [{ accessCookie: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, organizationId], properties: { name: { type: string }, description: { type: string }, organizationId: { type: string } } } } } }, responses: { '201': { description: Created }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } } }
...
  /environments:
    get: { security: [{ accessCookie: [] }], parameters: [{ name: projectId, in: query, required: true, schema: { type: string } }], responses: { '200': { description: Environment list }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } } }
    post: { security: [{ accessCookie: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, projectId], properties: { name: { type: string }, description: { type: string }, projectId: { type: string } } } } } }, responses: { '201': { description: Created }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } } }
  /secrets:
    get: { security: [{ accessCookie: [] }], parameters: [{ name: environmentId, in: query, required: true, schema: { type: string } }], responses: { '200': { description: Metadata only; never plaintext or encryption internals }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } } }
    post: { security: [{ accessCookie: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [key, value, environmentId], properties: { key: { type: string }, value: { type: string, format: password }, description: { type: string }, environmentId: { type: string } } } } } }, responses: { '201': { description: Created without encryption internals }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } } }
...